+++ src/routes/applet/dot_product/diagonal_parallelogram/StaticImage.svelte
